requirements.txt
azure-eventhub==5.11.4
python-dotenv==1.0.0
kafka-python==2.0.2
pandas==2.1.4
numpy==1.26.2
prometheus-client==0.19.0 